Refactor menu into MVC
