+++ src/main/resources/fr/iut/groupe/terraria/demo/vue/fenetre/MenuPrincipal.fxml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Text?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.image.ImageView?>
+
+<StackPane xmlns:fx="http://javafx.com/fxml/1" fx:controller="fr.iut.groupe.terraria.demo.controleur.interfacefx.MenuPrincipalController">
+    <children>
+        <ImageView fx:id="backgroundView" fitWidth="1920" fitHeight="1080" preserveRatio="false" />
+        <VBox alignment="TOP_CENTER" translateY="100">
+            <children>
+                <Text fx:id="titre" text="JUNGLE QUEST" styleClass="menu-title" />
+            </children>
+        </VBox>
+        <VBox fx:id="boutons" alignment="CENTER" spacing="60" styleClass="menu-boutons">
+            <children>
+                <Button fx:id="jouerButton" text="JOUER" styleClass="menu-button" />
+                <Button fx:id="quitterButton" text="QUITTER" styleClass="menu-button" />
+            </children>
+        </VBox>
+    </children>
+</StackPane>
